[mesh][metrics] split mesh metrics exporter setup - Move Prometheus exporter initialization into mesh_metrics module - Keep scrape listener, duration buckets, and upkeep timeout behavior compatible - Relocate exporter and Prometheus config tests to the new metrics modules
